requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.4
django-ckeditor==5.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
get==2019.4.13
gunicorn==20.0.4
Pillow==7.1.2
post==2019.4.13
psycopg2==2.8.5
public==2019.4.13
pytz==2020.1
query-string==2019.4.13
request==2019.4.13
sqlparse==0.3.1
whitenoise==5.1.0
